INT: updated api fetch call functions to make them reusable
